abbreviate the second and subsequent posts on the home page etc.
this is like loathhsxome: the first posting is included in full on any
page where it's included, but subsequent postings are abbreviated at the
first formfeed in the source, with a "more" link. unlike loathsxome,
the feed includes full text for all postings.

the obvious way to implement that was blocked by tidy, which removes the
formfeeds while parsing, so I had to do it in post.cpp.